Refactoring installation of demos
The demos were installed by creating files in user directories.

This is now changed to install files in the directory /opt/share/qt-demos.
Then during user startup, a service is run by systemd (preinstall-qt-demos)
that checks wether demos are installed or not. It installs the demo if
needed.

This patch improves a little the packaging of the demos but work remains to
be done: removing hard coded path to /opt/share and replace it with
build variables; improve the configuration of the settings.

Change-Id: Ice5d11e8d50aaccaa94995b2680cdfc273d27664
Signed-off-by: José Bollo <jose.bollo@open.eurogiciel.org>